% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_vector.R
\name{bradleyterry_from_sparse}
\alias{bradleyterry_from_sparse}
\title{Indicator matrix from two sparse matrices for Bradley-Terry models}
\usage{
bradleyterry_from_sparse(winners, losers)
}
\arguments{
\item{winners}{sparse matrix containing the winner for each observation}

\item{losers}{sparse matrix containing the loser for each observation}
}
\value{
sparse Bradley-Terry indicator matrix of type Matrix::dgCMatrix
}
\description{
Provides a helper function for generating indicator matrices
}
\examples{
winners <- Matrix::fac2sparse(c("k", "k", "l", "m", "m"))
losers <- Matrix::fac2sparse(c("l", "m", "m", "k", "l"))
Wwl <- bradleyterry_from_sparse(winners, losers)
}
