% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_vector.R
\name{bradleyterry_from_vectors}
\alias{bradleyterry_from_vectors}
\title{Indicator matrix from two vectors for Bradley-Terry models}
\usage{
bradleyterry_from_vectors(winners, losers)
}
\arguments{
\item{winners}{vector of strings containing the winner for each observation}

\item{losers}{vector of strings containing the loser for each observation}
}
\value{
sparse Bradley-Terry indicator matrix of type Matrix::dgCMatrix
}
\description{
Provides a helper function for generating indicator matrices
}
\examples{
winners <- c("k", "k", "l", "m", "m")
losers <- c("l", "m", "m", "k", "l")
Wwl <- bradleyterry_from_vectors(winners, losers)
}
