% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_vector.R
\name{bradleyterry_from_vectors2}
\alias{bradleyterry_from_vectors2}
\title{Indicator matrix from vectors for Bradley-Terry models}
\usage{
bradleyterry_from_vectors2(outcomes, a, b)
}
\arguments{
\item{outcomes}{win/lose outcome from the perspective of a, the "home team"}

\item{a}{vector of strings containing the "home team"}

\item{b}{vector of strings containing the "opposing team"}
}
\value{
sparse Bradley-Terry indicator matrix of type Matrix::dgCMatrix
}
\description{
Provides a helper function for generating indicator matrices
}
\examples{
outcomes <- rbinom(1, 5, .5)
a <- c("k", "k", "l", "m", "m")
b <- c("l", "m", "m", "k", "l")
W <- bradleyterry_from_vectors2(outcomes, a, b)
}
