% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{glmer_multimember}
\alias{glmer_multimember}
\title{Generalized linear mixed effects model with multimembership random effects}
\usage{
glmer_multimember(
  formula,
  data = NULL,
  family = gaussian,
  control = lme4::glmerControl(),
  start = NULL,
  verbose = 0L,
  nAGQ = 1L,
  subset,
  weights,
  na.action,
  offset,
  contrasts = NULL,
  mustart,
  etastart,
  devFunOnly = FALSE,
  memberships = list()
)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{data frame (possibly but not necessarily containing factors}

\item{memberships}{list of weights matrices  with which to replace Zt components}
}
\value{
lme4 model object
}
\description{
Version of lme4::glmer that allow for multimembership random effects
}
\examples{
df <- data.frame(
  x = runif(60, 0, 1),
  y = rbinom(60, 1, 0.6),
  memberships = rep(c("a,b,c", "a,c", "a", "b", "b,a", "b,c,a"), 10)
)
weights <- weights_from_vector(df$memberships)
glmer_multimember(y ~ x + (1|members), df, family=binomial, memberships=list(members=weights))
}
