% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\docType{class}
\name{glmerModMultiMember-class}
\alias{glmerModMultiMember-class}
\alias{glmerModMultiMember}
\title{Model object for multimembership generalized linear mixed models}
\value{
An object of class \code{glmerModMultiMember} similar to
\code{merModMultiMember} objects but inheriting from \code{glmerMod}
}
\description{
The \code{glmerModMultiMember} class extends \code{glmerMod}
from the \pkg{lme4}-package and \code{merModMultiMember}.
}
\seealso{
\code{\link[lme4]{glmer}} and \code{\link[lme4]{merMod}}
}
