% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_vector.R
\name{interaction_weights}
\alias{interaction_weights}
\title{Weight matrix from the interaction of two other weight matrices}
\usage{
interaction_weights(a, b)
}
\arguments{
\item{a}{sparse weight/indicator matrix of class Matrix::dgCmatrix}

\item{b}{sparse weight/indicator matrix of class Matrix::dgCmatrix}
}
\value{
sparse interaction weight/indicator matrix of class Matrix::dgCmatrix
}
\description{
Provides a helper function for pre-generating interactions.
Takes sparse weight/indicator matrices generated with e.g.
Matrix::fac2sparse() or weights_from_vector() as input.
}
\examples{
a <- rep(c("k", "l", "l,k"), 2)
b <- rep(c("m", "n"), 3)
Wa <- weights_from_vector(a)
Wb <- Matrix::fac2sparse(b)
Wab <- interaction_weights(Wa, Wb)
}
