% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{lmer}
\alias{lmer}
\title{Multimembership random effects}
\usage{
lmer(
  formula,
  data = NULL,
  REML = TRUE,
  control = lme4::lmerControl(),
  start = NULL,
  verbose = 0L,
  weights = NULL,
  na.action = na.omit,
  offset = NULL,
  contrasts = NULL,
  devFunOnly = FALSE,
  memberships = NULL
)
}
\arguments{
\item{formula}{mixed model formula}

\item{memberships}{list of weights matrices  with which to replace Zt components}

\item{...}{Other arguments as in \code{\link[lme4:lmer]{lme4::lmer()}}}
}
\value{
lme4 model object
}
\description{
Provides a wrapper around lme4::lmer to allow for multimembership random effects
}
\examples{
df <- data.frame(
  x = seq(60) + runif(60, 0, 10),
  y = seq(60) + rep(runif(6, 0, 10), 10),
  memberships = rep(c("a,b,c", "a,c", "a", "b", "b,a", "b,c,a"), 10)
)
weights <- weights_from_vector(df$memberships)
# note that the grouping variable name is arbitrary -- it just has
# to match the name in the list and doesn't need to correspond to a column name
# in the data
lmer(y ~ x + (1|members), df, memberships=list(members=weights))
}
