% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\docType{class}
\name{lmerModLmerTestMultiMember-class}
\alias{lmerModLmerTestMultiMember-class}
\alias{lmerModLmerTestMultiMember}
\title{Model object for multimembership linear mixed models with lmerTest}
\value{
An object of class \code{lmerModMultiMember} similar to
\code{merModMultiMember} objects but inheriting from \code{lmerMod}
}
\description{
The \code{lmerModMultiMember} class extends
\code{lmerModLmerTest} from the \pkg{lmerTest}-package and
\code{lmerModMultiMember}.
}
\seealso{
\code{\link[lmerTest]{lmer}} and \code{\link[lme4]{lmer}}
}
