% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\docType{class}
\name{lmerModMultiMember-class}
\alias{lmerModMultiMember-class}
\alias{lmerModMultiMember}
\title{Model object for multimembership linear mixed models}
\value{
An object of class \code{lmerModMultiMember} similar to
\code{merModMultiMember} objects but inheriting from \code{lmerMod}
}
\description{
The \code{lmerModMultiMember} class extends \code{lmerMod}
from the \pkg{lme4}-package and \code{merModMultiMember}.
}
\seealso{
\code{\link[lme4]{lmer}} and \code{\link[lme4]{merMod}}
}
