% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{plot_membership_hist.dgCMatrix}
\alias{plot_membership_hist.dgCMatrix}
\title{Plotting function for multiple membership weights histogram}
\usage{
\method{plot_membership_hist}{dgCMatrix}(x, varname = NULL, ...)
}
\arguments{
\item{x}{multimembership weight matrix}

\item{varname}{name of the multimembership variable, will be used as title}

\item{...}{additional arguments to pass on to graphics::hist}
}
\description{
Plotting function for multiple membership weights histogram
}
