% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{plot_membership_hist.merModMultiMember}
\alias{plot_membership_hist.merModMultiMember}
\title{Plotting method for multiple memberships model weights histogram}
\usage{
\method{plot_membership_hist}{merModMultiMember}(x, varname, ...)
}
\arguments{
\item{x}{merModMultiMember object}

\item{varname}{multimembership variable to be plotted}

\item{...}{additional arguments to pass on to graphics::hist}
}
\description{
Plotting method for multiple memberships model weights histogram
}
