% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{plot_membership_hist}
\alias{plot_membership_hist}
\title{Generic for multiple dispatch of multiple membership weights
histogram functions}
\usage{
plot_membership_hist(x, varname = NULL, ...)
}
\arguments{
\item{x}{object containing multimembership information, either a weight
matrix, multimembership model, or model summary}

\item{varname}{the multimembership variable to plot, if the object contains
more than one}

\item{...}{additional arguments to pass on to graphics::hist}
}
\description{
Generic for multiple dispatch of multiple membership weights
histogram functions
}
