% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{plot_membership_hist.summary.merModMultiMember}
\alias{plot_membership_hist.summary.merModMultiMember}
\title{Plotting method for multiple memberships model summary weights
histogram}
\usage{
\method{plot_membership_hist}{summary.merModMultiMember}(x, varname, ...)
}
\arguments{
\item{x}{merModMultiMember summary object}

\item{varname}{multimembership variable to be plotted}

\item{...}{additional arguments to pass on to graphics::hist}
}
\description{
Plotting method for multiple memberships model summary weights
histogram
}
