% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{plot_membership_matrix.dgCMatrix}
\alias{plot_membership_matrix.dgCMatrix}
\title{Plotting function for multiple membership matrix}
\usage{
\method{plot_membership_matrix}{dgCMatrix}(x, varname = NULL, ...)
}
\arguments{
\item{x}{multimembership weight matrix}

\item{varname}{name of the multimembership variable, will be used as title}

\item{...}{additional arguments to pass on to Matrix::image}
}
\description{
Plotting function for multiple membership matrix
}
