% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{plot_membership_matrix.merModMultiMember}
\alias{plot_membership_matrix.merModMultiMember}
\title{Plotting method for multiple memberships model}
\usage{
\method{plot_membership_matrix}{merModMultiMember}(x, varname)
}
\arguments{
\item{x}{merModMultiMember object}

\item{varname}{multimembership variable to be plotted}
}
\description{
Plotting method for multiple memberships model
}
