% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{plot_membership_matrix}
\alias{plot_membership_matrix}
\title{Generic for multiple dispatch of multiple membership plot functions}
\usage{
plot_membership_matrix(x, varname = NULL)
}
\arguments{
\item{x}{object containing multimembership information, either a weight
matrix, multimembership model, or model summary}

\item{varname}{the multimembership variable to plot, if the object contains
more than one}
}
\description{
Generic for multiple dispatch of multiple membership plot functions
}
