% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{print.summary.lmerModMultiMember}
\alias{print.summary.lmerModMultiMember}
\title{print method for multiple memberships summary for lmer}
\usage{
\method{print}{summary.lmerModMultiMember}(x, ...)
}
\arguments{
\item{x}{lmerModMultiMember object}

\item{...}{additional arguments to be passed on to print.summary.merMod}
}
\description{
print method for multiple memberships summary for lmer
}
