% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{print.summary.merModMultiMember}
\alias{print.summary.merModMultiMember}
\title{Print method for multiple memberships model summary}
\usage{
\method{print}{summary.merModMultiMember}(x, ...)
}
\arguments{
\item{x}{merModMultiMember object}

\item{...}{additional arguments to be passed on to print.summary.merMod}
}
\description{
Print method for multiple memberships model summary
}
