% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{summary.glmerModMultiMember}
\alias{summary.glmerModMultiMember}
\title{summary method for multimembership model objects}
\usage{
\method{summary}{glmerModMultiMember}(object, ...)
}
\arguments{
\item{object}{glmerModMultiMember model object}

\item{...}{additional arguments to be passed on to summary.glmerMod}
}
\value{
summary of glmerModMultiMember object
}
\description{
summary method for multimembership model objects
}
