% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{summary.lmerModMultiMember}
\alias{summary.lmerModMultiMember}
\title{summary method for multimembership model objects}
\usage{
\method{summary}{lmerModMultiMember}(object, ...)
}
\description{
summary method for multimembership model objects
}
