% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{summary.merModMultiMember}
\alias{summary.merModMultiMember}
\title{Summary method for multimembership model objects}
\usage{
\method{summary}{merModMultiMember}(object, ...)
}
\arguments{
\item{object}{merModMultiMember model object}

\item{...}{additional arguments to be passed on to summary.merMod}
}
\value{
summary of merModMultiMember object
}
\description{
Summary method for multimembership model objects
}
