% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_vector.R
\name{weights_from_columns}
\alias{weights_from_columns}
\title{Weight matrix from columns containing memberships}
\usage{
weights_from_columns(membership_columns)
}
\arguments{
\item{membership_columns}{columns containing group memberships}
}
\value{
sparse weight matrix
}
\description{
Provides a helper function for generating weight matrices
}
\examples{
member_cols <- cbind(
  c("a", "a", "b", NA, "c"),
  c("b", "c", "a", "c", "a"),
  c("c", NA, "c", NA, NA)
)
weights_from_columns(member_cols)
}
