% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_vector.R
\name{weights_from_vector}
\alias{weights_from_vector}
\title{Weight matrix from vector of memberships}
\usage{
weights_from_vector(membership_vector, sep = ",")
}
\arguments{
\item{membership_vector}{vector of strings containing group memberships}

\item{sep}{separator that delimits group memberships in membership_vector}
}
\value{
sparse weight/indicator matrix of type Matrix::dgCmatrix
}
\description{
Provides a helper function for generating weight matrices
}
\examples{
a <- c("k,l,m", "k,m", "k", "l", "l,k")
Wa <- weights_from_vector(a)
}
