% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{glmer}
\alias{glmer}
\title{Multimembership random effects}
\usage{
glmer(
  formula,
  data = NULL,
  family,
  control = lme4::glmerControl(),
  start = NULL,
  verbose = 0L,
  nAGQ = 1L,
  weights = NULL,
  na.action = na.omit,
  offset = NULL,
  contrasts = NULL,
  devFunOnly = FALSE,
  memberships = NULL
)
}
\arguments{
\item{formula}{mixed model formula}

\item{memberships}{list of weights matrices  with which to replace Zt components}

\item{...}{Other arguments as in \code{\link[lme4:glmer]{lme4::glmer()}}}
}
\value{
lme4 model object
}
\description{
Provides a wrapper around lme4::glmer to allow for multimembership random effects
}
\examples{
df <- data.frame(
  x = runif(60, 0, 1),
  y = rbinom(60, 1, 0.6),
  memberships = rep(c("a,b,c", "a,c", "a", "b", "b,a", "b,c,a"), 10)
)
weights <- weights_from_vector(df$memberships)
# note that the grouping variable name is arbitrary -- it just has
# to match the name in the list and doesn't need to correspond to a column name
# in the data
glmer(y ~ x + (1|members), df, family=binomial, memberships=list(members=weights))
}
