% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{lmer_multimember}
\alias{lmer_multimember}
\title{Multimembership random effects}
\usage{
lmer_multimember(formula, data, memb_mat = list(), ...)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{data frame (possibly but not necessarily containing factors}

\item{memb_mat}{list of weights matrices  with which to replace Zt components}

\item{...}{additional arguments to pass through to lme4::lmer}
}
\value{
lme4 model object
}
\description{
Provides a wrapper around lme4::lmer to allow for multimembership random effects
}
\examples{
df <- data.frame(
  x = seq(5),
  y = seq(5),
  memberships = c("a,b,c", "a,c", "a", "b", "b,a")
)
weights <- weights_from_vector(df$memberships)
lmer_multimember(y ~ x + (1|members), df, memb_mat=list(members=weights))
}
