% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{summary.lmerModMultiMember}
\alias{summary.lmerModMultiMember}
\title{summary method for multimembership model objects}
\usage{
\method{summary}{lmerModMultiMember}(object, ...)
}
\arguments{
\item{object}{lmerModMultiMember model object}

\item{...}{additional arguments to be passed on to summary.lmerMod}
}
\value{
summary of lmerModMultiMember object
}
\description{
summary method for multimembership model objects
}
