% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_vector.R
\name{weights_from_columns}
\alias{weights_from_columns}
\title{Weight matrix from columns containing memberships}
\usage{
weights_from_columns(membership_columns)
}
\arguments{
\item{membership_columns}{columns containing group memberships}
}
\value{
sparse weight/indicator matrix of type Matrix::dgCmatrix
}
\description{
Provides a helper function for generating weight matrices
}
\examples{
a <- cbind(
  c("k", "k", "l", NA, "m"),
  c("l", "m", "k", "m", "k"),
  c("m", NA, "m", NA, NA)
)
Wa <- weights_from_columns(a)
}
