% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_multimember.R
\name{weights_from_vector}
\alias{weights_from_vector}
\title{Weight matrix from vector of memberships}
\usage{
weights_from_vector(membership_vector, sep = ",")
}
\arguments{
\item{membership_vector}{vector of strings containing group memberships}

\item{sep}{separator that delimits group memberships in membership_vector}
}
\value{
sparse weight matrix
}
\description{
Provides a helper function for generating weight matrices
}
\examples{
member_vec <-  c("a,b,c", "a,c", "a", "b", "b,a")
weights_from_vector(member_vec)
}
